requirements.txt
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
cffi==1.11.5
channels==2.1.3
channels-redis==2.3.0
constantly==15.1.0
cryptography==2.2.2
daphne==2.2.2
diff-match-patch==20121119
Django==2.0.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-reversion==3.0.0
et-xmlfile==1.0.1
future==0.15.2
hiredis==0.2.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
msgpack==0.5.6
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pycparser==2.18
PyHamcrest==1.9.0
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
Twisted==18.7.0
txaio==18.8.1
unicodecsv==0.14.1
uWSGI==2.0.17.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
